# api/openapi/console.yaml
...
openapi: 3.1.0
info:
  title: Ecoma LLM Gateway — Console API
  version: 0.1.0
  description: >-
    The Control Plane's public surface, and the only one of the gateway's
    three contracts a browser reaches. Every response returns X-Request-Id,
    and every failure — including the ones no operation declares, a 404 for
    an unmatched path and a 500 for an unclassifiable implementation failure
    — returns the ErrorEnvelope. A readiness probe taken while the Control
    Plane's own dependencies are not answering returns HTTP 503 and the
    `service_unavailable` code.

    The surface has two halves that meet at the credential on the request. The
    **session surface** — sign-in, the current principal, sign-out — is the
    only part reachable without one, and the only part that writes nothing but
    the session itself. The **product surface** is everything else, and every
    operation on it authorizes against the account the request's credential
    names. That credential is the session cookie for all of them but one: the
    usage operation accepts `consoleCredential` instead, a bearer credential for
    the caller that has no browser at all — a script, a scheduled report, an
    operator's tooling — because a session is a thing a browser holds and this
    surface is not only read by browsers.

    The session credential is declared as `sessionCookie` in
    `components.securitySchemes` and named by every operation that requires
...

...

...

...

...
servers:
  - url: /
...
paths:
...
  /top-up-offers:
    get:
      operationId: listTopUpOffers
      tags:
        - payments
      summary: The top-up offers this deployment sells
      description: >-
        What a customer may choose to top up with: each offer's identifier, the
        amount it charges, the currency that amount is in, and a label to
        render. It is the price list, and a console that rendered a top-up
        chooser without it would be asking a customer to name an offer they
        could not see — a blind POST, which is not a screen.

        This is the ONLY operation on this surface that publishes a price, and
        publishing it read-only is what keeps the design whole rather than
        breaking it. The request side still names an OFFER and never an amount:
        `CreatePaymentIntentRequest` carries no `amount_minor_units` and no
        `currency`, and nothing here changes that. Showing a price and being
        able to set one are different powers, and they are held by different
        operations — a client that reads these figures is reading what the
        server charges, not choosing it.

        The offers are this DEPLOYMENT's own configuration rather than
        anybody's rows: a price this plane decides, not a fact about an
        account. That is why the operation is not account-scoped, carries no
        account parameter, and returns the same list to every session that may
        fund itself.

        The response is deliberately not a `PageEnvelope`, and the reason is
        the one this repository already gives for rejecting page numbers: a
        cursor names a POSITION, and a fixed price list has no position to
        name. There is no `total` either — a count maintained for a list that
        arrives whole would be a number that goes stale for nobody's benefit.
      security:
        - sessionCookie: []
      parameters:
        - $ref: "./shared/probes.yaml#/components/parameters/RequestId"
      responses:
        "200":
          description: The offers this deployment publishes.
          headers:
            X-Request-Id:
              $ref: "./shared/console.yaml#/components/headers/RequestID"
            Cache-Control:
              $ref: "./shared/console.yaml#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "./shared/console.yaml#/components/schemas/TopUpOfferList"
        "401":
          $ref: "./shared/console.yaml#/components/responses/Unauthenticated"
        "404":
          $ref: "./shared/errors.yaml#/components/responses/NotFound"
        "405":
          $ref: "./shared/errors.yaml#/components/responses/MethodNotAllowed"
        "500":
          $ref: "./shared/errors.yaml#/components/responses/InternalError"
...
components:
  securitySchemes:
    consoleCredential:
      type: http
      scheme: bearer
      description: >-
        A credential presented by a caller of the Control Plane's public
        surface. It is the one scheme here that is not a browser session: the
        operations beside this one identify their caller by the session
        cookie, and this is for a caller that has none — a script, a scheduled
        report, an operator's tooling.

        The deployment supplies the credential and the resolution is the
        server's: the account this surface answers about is derived from the
        credential and is never named by the request. A credential that does
        not resolve is HTTP 401 with `unauthenticated`; it is not silently
        answered for a default
        account, because a caller that cannot tell it was refused is a
        caller that will draw a chart of someone else's spend.

        The transport is expected to be replaced by one that proves the
        identity cryptographically (mTLS or a signed credential) without
        changing this contract, and a browser session is not this scheme: a
        session is a different credential with a different lifetime and a
        different revocation story, and modelling it here before it exists
        would be a promise this document cannot keep.
        "405":
          $ref: "./shared/errors.yaml#/components/responses/MethodNotAllowed"
        "500":
          description: >-
            Nothing was durably recorded, and the code is `internal`. This is
            the ONE case in which a redelivery can change the outcome, and it
            is the only 5xx here for that reason: the delivery is not on file,
            no state moved, and the provider's next attempt is the correct
            next step rather than a repetition of a decision already made.
            Internal causes are logged against the request identifier and
            never serialized.
          headers:
            X-Request-Id:
              $ref: "./shared/console.yaml#/components/headers/RequestID"
            Cache-Control:
              $ref: "./shared/console.yaml#/components/headers/NoStore"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-console_session
      description: >-
        The console's session, presented by the browser that holds it: the
        only credential this surface accepts, and a cookie rather than a
        bearer token because the caller population is browsers. The browser
        attaches it to every same-origin request by itself, no page script
        ever reads its value — it is `HttpOnly` — and a credential that cannot
        be read is a credential that cannot be pasted into a request built
        somewhere else. Its value is an opaque 256-bit identifier, the session
        row it names stores only the SHA-256 digest of it, and the `Secure`,
        `SameSite=Strict` and `Path=/` attributes are part of the credential
        rather than conveniences around it; the `SessionCookie` header states
        the same value from the response's side.

        An operation that requires this credential names it in its own
        `security`, and an operation that requires none says `security: []`.
        The empty list is not a gap left for a reader to fill in: it is the
        statement that no cookie is asked for on that row, and the difference
        matters in both directions. Declared the other way round, this scheme
        would tell a generated client that sign-in needs a session — the one
        operation that creates one — that sign-out needs a session a caller
        has just lost, and that a probe needs a session an orchestrator has
        never had.

        It is not the only guard on a write, and it is the only SCHEME here.
        Every unsafe operation additionally clears a same-origin `Origin`, a
        JSON content type and a double-submit header, and none of those is a
        second security scheme because none of them is a credential: each is
        a check that a browser made this request rather than a script on
        another site, they are enforced together and cannot be satisfied
        separately, and a client that could choose among them would be a
        client choosing which half of its own CSRF protection to perform. The
        webhook at `POST /payment-webhooks/{provider}` declares `security: []`
        as well, for the one reason a row on this surface can: it is
        authenticated by a signature over the request body that only the
        provider's servers and this deployment can compute, so no cookie is
        asked for and none would help.